refactor(posters_import,utils): move helper functions to utils.py
Move a number of helper/utility functions used in
management command for importing posters to new
`utils.py` file.
